+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Version latest (0.5.0?, 2019-09-24?)
 - ...
 - enh: display an instance's tags in the optimization table
 - feat: add `--filter-tags` option
+- enh: display n ec2 analysed
+- enh: move out the call to listeners from inside `handle_ec2_obj`
 
 
 Version 0.5.0 (2019-09-25)