+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Version latest (0.5.0?, 2019-09-24?)
 Version 0.5.0 (2019-09-25)
 
 - enh: readme docs for examples was not up-to-date with monthly savings in `--optimize`. Updated.
-- feat: add datadog integration
+- feat: merge datadog integration
 - bugfix: moved the try/finally from the `__main__` section in `cli.py` to the `cli()` function
 - enh: moved test for datadogManager to separate file
 - bugfix: querying from datadog for a single host was missing the `host:` prefix